cape: Configure ramoops at a static address in RAM
Starting with SM8450, Qualcomm platforms no longer support normal warm
reboots. As a result, it's not possible for the HLOS kernel to trigger a
warm reboot that'll preserve the contents of DRAM (and thus preserve a
ramoops); instead, XBL_SC always runs through the cold boot flow for DRAM,
with only two exceptions: full ramdumps and minidumps.

In XBL_SC, which isn't customizable by OEMs starting with SM8350, the warm
boot flow for DRAM will only be honored if:

 1. PBL indicates that the device is booting into EDL. This is indicated to
    XBL_SC via IMEM.

 2. PSHOLD is configured to perform a warm reset.

 3. The TCSR register contains either QCOM_DOWNLOAD_FULLDUMP (0x10) or
    QCOM_DOWNLOAD_MINIDUMP (0x20); i.e., `(TCSR & 0x30) != 0` is the
    condition that must be fulfilled.

    3a. A Debug Policy (DP) image flashed to the apdp partition must have
        a crash-dump mode specified: either NS_FULL_DUMP/SEC_FULL_DUMP in
        order to permit writing 0x10 to the TCSR register, or MINI_DUMP to
        permit writing 0x20 to the TCSR register. Without the DP image, any
        attempt to write 0x10 or 0x20 to the TCSR register won't latch.

So, in order to perform a warm reboot for the sake of collecting a ramoops,
the only option is to reboot the device into crash-dump mode, which boots
into XblRamdump.elf. Once booted into XblRamdump.elf, everything in DRAM
will still be intact since XBL_SC will have run through the warm boot flow
for DRAM. Thus, once inside XblRamdump.elf, we copy the contents of the
pstore buffer to the logdump partition in UFS so that it can persist, and
then trigger a cold reboot to get back into Android. Just before Linux is
booted, ABL reads out the contents of the pstore buffer that were saved to
UFS and copies it into the exact address in DRAM where the pstore buffer
resided, thereby keeping a ramoops intact after a reboot. Then when ABL
jumps to Linux, Linux observes that the ramoops from the previous boot is
located intact in DRAM exactly where it's expected.

This scheme requires that Linux, ABL, and XblRamdump.elf all agree upon the
address and size of the pstore buffer in DRAM. Furthermore, ABL and
XblRamdump.elf must agree upon the partition in UFS where the pstore buffer
is copied.

Configure ramoops to use a free (as per /proc/iomem) static address to make
this possible, and double the buffer sizes while we're at it.

Change-Id: I306ab1bc57fff703d188ededfbabca161b444cb4
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>